Move `OutputStream`/`InputStream` definitions outside of the headers
